refactor: address handling - wip: ensureReedSolomonAddress
