+++ lib/packages/core/src/address/index.ts
@@ -1,2 +1,2 @@
-export * from './address';
+export {Address} from './address';
